requirements.txt
fastapi~=0.115.2
starlette~=0.40.0
uvicorn~=0.15.0
jwt~=1.3.1
PyJWT~=2.9.0
python-dotenv~=1.0.1